% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nadda_functionized.R
\name{count_kmers}
\alias{count_kmers}
\title{Counting k-mers in the dataset.}
\usage{
count_kmers(obj, klen = 6, parallel = TRUE, nproc = ifelse(parallel,
  pbdMPI::comm.size(), 1), distributed = FALSE)
}
\arguments{
\item{obj}{A filepath to a fasta file containing protein sequences or an 
AAStringSet object containing the sequences}

\item{klen}{length of the k-mers to be used}

\item{parallel}{Indicating whether the operation should be p
erformed in parallel}

\item{nproc}{Currently not supported. Will use all processors 
available to the job on cluster}

\item{distributed}{A boolean, indicating whether the data is 
spread among multiple processors.}
}
\value{
Returns a dataframe with two columns. Each row includes 
one k-mer and an integer indicating the number of 
times that k-mer appears in the input dataset. Each k-mer in 
each sequence is counted
at most once, i.e., if there are multiple occureneces of a 
k-mer in one sequence, only one of them is counted.
}
\description{
counts the number of times each k-mer appears in the 
dataset and returns a dataframe indicating 
these counts.
Each k-mer in each sequence is counted
at most once, i.e., if there are multiple occureneces of a k-mer in one 
sequence, only one of them is counted.
}
\details{
If \emph{parallel} is set to \strong{TRUE} and  
\emph{distributed} is set to \strong{FALSE}, the method 
distributes the data between different 
processors and sets \emph{distributed} to \strong{TRUE}. 
Otherwise, if the \emph{parallel} is set to \strong{FALSE} 
and \emph{distributed} is set to \strong{TRUE}, 
the kmer frequencies are computed on each processor separately 
but then communicated between each other, and therefore 
at the end all processors have the same set of frequencies 
for kmers stored, using which they will generate frequency 
profiles for their chunk of sequences.
If you prefer to run the operation in serial, set both 
\emph{parallel} and \emph{distributed} to \strong{FALSE}.
}
\examples{
library(Biostrings)
library(data.table)
## Generate a set of three example protein sequences
seqs <- AAStringSet(c("seq1"="MLVVD",
                      "seq2"="PVVRA",
                      "seq3"="LVVR"))
## Count the kmers and generate a dataframe of the frequencies
str(seqs)
length(seqs)
freqs <- count_kmers(seqs, klen = 3, parallel = FALSE)
head(freqs)
##    kmer count
##1:  LVV  2
##2:  MLV  1
##3:  PVV  1
##4:  VRA  1
##5:  VVD  1
##6:  VVR  2
}
\author{
Armen Abnousi
}
